inlineXML stored/avail as string + contentresolver copes with relative refs
The inlineXML change fixes a potential issue reported by findbugs, but
also results in a better design -- it makes DatastreamVersions more
convenient to work with and obviates the need to do the conversion
on each FedoraDTO.toString() call.

+++ fcrepo-dto/fcrepo-dto-core/src/main/java/com/github/cwilper/fcrepo/dto/core/io/DefaultContentHandler.java
@@ -61,7 +61,7 @@ public class DefaultContentHandler implements ContentHandler {
         } else {
             // if pid isn't assigned yet, come up with a unique enough path
             // to avoid collisions for now
-            sb.append("nopid:" + Math.abs(obj.hashCode()));
+            sb.append("nopid:" + Math.abs((long) obj.hashCode()));
         }
         sb.append("/");
         sb.append(ds.id());
